fix: get all song respone and duration type

+++ migrations/1706588646193_create-table-song.js
@@ -25,7 +25,7 @@ exports.up = pgm => {
       notNull: true,
     },
     duration: {
-      type: 'BIGINT',
+      type: 'INTEGER',
       notNull: false,
     },
     album_id: {
